Revert "consolidated trainer settings in single file"
This reverts commit e8f26512ac8ac3f03fef039b6eaf68b00462ba9c.
